Better font error handling, Creature Text path fix
I don't know why I was being so case sensitive with the previous method of grabbing the correct path info, but it SHOULD be fixed. Also it's messy with a bunch of try excepts but this way I don't have to futz around with OS paths and finding installed fonts that way. Too machine dependent.