Proxyshop 1.4 compatibility update
A pretty sweeping update with a few tweaks in logic, but largely working as intended. Invocation is a work in progress and does not meet my personal standards, it's mostly just a repurposing of old CardConjurer assets and testing that pipeline translating into proxyshop.

Be sure to grab the new .psd files as well.